+++ src/providers/Completion/GMdotAPIsCompletionProvider.ts
@@ -20,7 +20,7 @@ const GMdotAPIsCompletionProvider = vscode.languages.registerCompletionItemProvi
                             doc.appendMarkdown(i.value);
                             break;
                         case 'code':
-                            doc.appendCodeblock(i.value, 'javascript');
+                            doc.appendCodeblock(i.value, 'typescript');
                             break;
                         default:
                             doc.appendText(i.value);