+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-- Nothing
+- Support medadata `@exclude`, `@exclude-match`.
 
 ## [0.0.1] - 2022-07-25